+++ src/components/top250-pagination.tsx
@@ -36,10 +36,10 @@ export function Top250Pagination({ page }: Top250PaginationProps) {
     return (
         <nav className="pagination">
             <a className={`pagination-previous ${page <= 1 ? 'do-hide' : ''}`} data-page={page - 1}>
-                上一页
+                上一頁
             </a>
             <a className={`pagination-next ${hasMore ? 'do-hide' : ''}`} data-page={page + 1}>
-                下一页
+                下一頁
             </a>
             <ul className="pagination-list">
                 {Array.from({ length: 5 }, (_, i) => i + 1).map((i) => (