+++ src/plugins/related-plugin.tsx
@@ -114,7 +114,7 @@ export class RelatedPlugin extends BasePlugin {
         try {
             list = await K(movieId, 1, limit);
         } catch (err: any) {
-            if (err.toString().includes('簽名已过期')) {
+            if (err.toString().includes('簽名已過期')) {
                 show.error('生成签名失败, 请检查系统时间及时区是否正确!');
             }
             clog.error('获取清单失败:', err);